requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
pillow==10.0.1
requests==2.31.0
dnspython==2.4.2 
aiohttp